+++ pacman-ui-fx-2d/src/main/java/de/amr/games/pacman/ui/fx/rendering2d/PacManGameRenderer.java
@@ -124,7 +124,7 @@ public class PacManGameRenderer implements Rendering2D {
 	}
 
 	@Override
-	public void drawMaze(GraphicsContext g, int x, int y, int mazeNumber, World world) {
+	public void drawMaze(GraphicsContext g, double x, double y, int mazeNumber, World world) {
 		var flashingAnimation = world.animation(GameModel.AK_MAZE_FLASHING);
 		if (flashingAnimation.isPresent() && flashingAnimation.get().isRunning()) {
 			var flashing = (boolean) flashingAnimation.get().frame();
